Reorganized Conditions into separate folder/namespace
